Data simulation for binary endpoints for a platform trial with an arbitrary number of treatment arms entering sequentially

Simulates data from a platform trial with binary endpoints and an arbitrary number of treatment arms entering sequentially. There is the option to either specify the timing of adding arms and the overall sample size in the trial or specify the sample size per arm (assumed equal) and the allocation ratios per period.

datasim_bin(
  n_total,
  num_arms,
  d,
  n_arm,
  alloc_ratios,
  period_blocks = 2,
  p0,
  OR,
  lambda,
  trend,
  N_peak,
  full = FALSE
)

Arguments

n_total

Overall sample size for the trial

num_arms

Number of treatment arms in the trial

d

Timing of adding new arms in terms of number of patients allocated to the control arm

n_arm

Sample size per arm

alloc_ratios

Matrix with allocation ratios in each period (rows = arms; columns = periods)

period_blocks

Number to multiply the number of active arms with in order to get the block size per period

p0

Response in the control arm

OR

Vector with odds ratios for each arm

lambda

Vector with strength of time trend in each arm

trend

Indicates the time trend pattern ("linear", "stepwise" or "inv_u")

N_peak

Point at which the inverted-u time trend switches direction in terms of overall sample size

full

Boolean. Indicates whether only variables needed for the analysis should be in the output (FALSE) or also additional information (lambdas, underlying responses) should be included (TRUE). Default=FALSE

Value

Data frame: simulated trial data

Author

Pavla Krotka, Marta Bofill Roig

Examples


# Usage option 1 - specify n_total, num_arms and d
datasim_bin(n_total = 1000, num_arms = 3, d = 120, n_arm = NULL, alloc_ratios = NULL, period_blocks = 2, p0 = 0.7, OR = rep(1.4, 3), lambda = rep(0.15, 4), trend = "linear")
#> Loading required package: rlang
#> Warning: package 'rlang' was built under R version 3.6.3
#>      response treatment period
#> 1           1         0      1
#> 2           1         0      1
#> 3           1         1      1
#> 4           1         1      1
#> 5           1         1      1
#> 6           1         0      1
#> 7           1         0      1
#> 8           1         1      1
#> 9           1         0      1
#> 10          1         0      1
#> 11          0         1      1
#> 12          1         1      1
#> 13          1         0      1
#> 14          1         1      1
#> 15          0         1      1
#> 16          1         0      1
#> 17          1         0      1
#> 18          1         1      1
#> 19          1         0      1
#> 20          1         1      1
#> 21          0         1      1
#> 22          1         1      1
#> 23          1         0      1
#> 24          1         0      1
#> 25          1         0      1
#> 26          1         1      1
#> 27          1         0      1
#> 28          1         1      1
#> 29          1         1      1
#> 30          1         0      1
#> 31          1         0      1
#> 32          0         1      1
#> 33          1         0      1
#> 34          1         0      1
#> 35          1         1      1
#> 36          1         1      1
#> 37          1         0      1
#> 38          1         1      1
#> 39          0         1      1
#> 40          1         0      1
#> 41          0         1      1
#> 42          1         0      1
#> 43          0         0      1
#> 44          0         1      1
#> 45          1         1      1
#> 46          1         0      1
#> 47          0         0      1
#> 48          1         1      1
#> 49          1         0      1
#> 50          0         1      1
#> 51          1         1      1
#> 52          1         0      1
#> 53          1         1      1
#> 54          1         0      1
#> 55          0         0      1
#> 56          1         1      1
#> 57          1         1      1
#> 58          1         0      1
#> 59          1         0      1
#> 60          1         1      1
#> 61          0         1      1
#> 62          1         0      1
#> 63          1         1      1
#> 64          1         0      1
#> 65          1         1      1
#> 66          1         0      1
#> 67          0         0      1
#> 68          0         1      1
#> 69          0         0      1
#> 70          1         0      1
#> 71          0         1      1
#> 72          1         1      1
#> 73          0         0      1
#> 74          1         1      1
#> 75          1         1      1
#> 76          1         0      1
#> 77          0         0      1
#> 78          1         0      1
#> 79          1         1      1
#> 80          1         1      1
#> 81          1         0      1
#> 82          0         1      1
#> 83          1         0      1
#> 84          1         1      1
#> 85          1         0      1
#> 86          0         0      1
#> 87          0         1      1
#> 88          1         1      1
#> 89          0         1      1
#> 90          0         0      1
#> 91          1         0      1
#> 92          1         1      1
#> 93          1         0      1
#> 94          1         0      1
#> 95          1         1      1
#> 96          1         1      1
#> 97          1         0      1
#> 98          1         1      1
#> 99          1         1      1
#> 100         1         0      1
#> 101         0         0      1
#> 102         1         1      1
#> 103         0         1      1
#> 104         0         0      1
#> 105         0         0      1
#> 106         1         0      1
#> 107         1         1      1
#> 108         0         1      1
#> 109         1         1      1
#> 110         0         0      1
#> 111         1         1      1
#> 112         1         0      1
#> 113         1         1      1
#> 114         1         1      1
#> 115         1         0      1
#> 116         0         0      1
#> 117         1         1      1
#> 118         0         0      1
#> 119         1         1      1
#> 120         0         0      1
#> 121         1         1      1
#> 122         0         0      1
#> 123         0         1      1
#> 124         0         0      1
#> 125         1         1      1
#> 126         1         0      1
#> 127         0         1      1
#> 128         1         0      1
#> 129         0         1      1
#> 130         1         1      1
#> 131         1         0      1
#> 132         1         0      1
#> 133         1         1      1
#> 134         0         0      1
#> 135         1         0      1
#> 136         0         1      1
#> 137         1         1      1
#> 138         1         1      1
#> 139         1         0      1
#> 140         1         0      1
#> 141         1         1      1
#> 142         0         0      1
#> 143         1         1      1
#> 144         1         0      1
#> 145         1         1      1
#> 146         1         0      1
#> 147         1         1      1
#> 148         1         0      1
#> 149         1         1      1
#> 150         1         1      1
#> 151         1         0      1
#> 152         1         0      1
#> 153         1         0      1
#> 154         1         1      1
#> 155         1         0      1
#> 156         1         1      1
#> 157         1         0      1
#> 158         0         0      1
#> 159         1         1      1
#> 160         1         1      1
#> 161         1         1      1
#> 162         1         1      1
#> 163         1         0      1
#> 164         1         0      1
#> 165         1         0      1
#> 166         1         1      1
#> 167         1         1      1
#> 168         1         0      1
#> 169         0         1      1
#> 170         1         1      1
#> 171         1         0      1
#> 172         0         0      1
#> 173         0         0      1
#> 174         1         1      1
#> 175         0         1      1
#> 176         1         0      1
#> 177         1         0      1
#> 178         1         1      1
#> 179         0         0      1
#> 180         0         1      1
#> 181         1         1      1
#> 182         0         1      1
#> 183         1         0      1
#> 184         1         0      1
#> 185         1         1      1
#> 186         1         1      1
#> 187         1         0      1
#> 188         1         0      1
#> 189         1         1      1
#> 190         1         0      1
#> 191         0         0      1
#> 192         1         1      1
#> 193         0         1      1
#> 194         1         1      1
#> 195         0         0      1
#> 196         1         0      1
#> 197         0         1      1
#> 198         1         0      1
#> 199         1         1      1
#> 200         1         0      1
#> 201         1         1      1
#> 202         1         0      1
#> 203         1         1      1
#> 204         1         0      1
#> 205         0         1      1
#> 206         1         0      1
#> 207         1         1      1
#> 208         1         0      1
#> 209         1         0      1
#> 210         0         0      1
#> 211         1         1      1
#> 212         1         1      1
#> 213         1         0      1
#> 214         1         1      1
#> 215         1         0      1
#> 216         1         1      1
#> 217         1         0      1
#> 218         1         1      1
#> 219         1         1      1
#> 220         1         0      1
#> 221         0         0      1
#> 222         1         1      1
#> 223         0         1      1
#> 224         1         0      1
#> 225         1         0      1
#> 226         1         1      1
#> 227         1         1      1
#> 228         1         0      1
#> 229         1         1      1
#> 230         1         0      1
#> 231         1         0      1
#> 232         1         1      1
#> 233         1         1      1
#> 234         1         1      1
#> 235         1         0      1
#> 236         1         0      1
#> 237         1         1      1
#> 238         1         0      1
#> 239         1         0      1
#> 240         0         1      1
#> 241         0         1      2
#> 242         1         2      2
#> 243         0         0      2
#> 244         0         1      2
#> 245         0         0      2
#> 246         0         2      2
#> 247         0         0      2
#> 248         1         1      2
#> 249         0         2      2
#> 250         1         0      2
#> 251         1         1      2
#> 252         0         2      2
#> 253         1         1      2
#> 254         1         2      2
#> 255         1         1      2
#> 256         1         2      2
#> 257         1         0      2
#> 258         1         0      2
#> 259         1         2      2
#> 260         0         0      2
#> 261         1         0      2
#> 262         1         2      2
#> 263         0         1      2
#> 264         1         1      2
#> 265         0         0      2
#> 266         1         2      2
#> 267         1         2      2
#> 268         1         0      2
#> 269         1         1      2
#> 270         1         1      2
#> 271         1         0      2
#> 272         0         1      2
#> 273         1         2      2
#> 274         1         1      2
#> 275         0         0      2
#> 276         1         2      2
#> 277         1         1      2
#> 278         1         0      2
#> 279         0         0      2
#> 280         1         2      2
#> 281         1         1      2
#> 282         0         2      2
#> 283         0         1      2
#> 284         1         0      2
#> 285         1         1      2
#> 286         1         0      2
#> 287         1         2      2
#> 288         1         2      2
#> 289         1         1      2
#> 290         1         2      2
#> 291         1         2      2
#> 292         1         0      2
#> 293         0         0      2
#> 294         1         1      2
#> 295         0         0      2
#> 296         1         1      2
#> 297         1         0      2
#> 298         0         2      2
#> 299         1         2      2
#> 300         1         1      2
#> 301         1         1      2
#> 302         1         0      2
#> 303         1         2      2
#> 304         0         0      2
#> 305         1         2      2
#> 306         1         1      2
#> 307         1         0      2
#> 308         0         0      2
#> 309         0         2      2
#> 310         1         2      2
#> 311         1         1      2
#> 312         1         1      2
#> 313         1         1      2
#> 314         1         2      2
#> 315         1         0      2
#> 316         1         1      2
#> 317         1         2      2
#> 318         0         0      2
#> 319         0         0      2
#> 320         0         1      2
#> 321         1         1      2
#> 322         0         2      2
#> 323         1         0      2
#> 324         1         2      2
#> 325         1         0      2
#> 326         1         2      2
#> 327         1         1      2
#> 328         1         0      2
#> 329         1         2      2
#> 330         0         1      2
#> 331         1         1      2
#> 332         1         0      2
#> 333         0         0      2
#> 334         1         1      2
#> 335         1         2      2
#> 336         1         2      2
#> 337         0         0      2
#> 338         1         2      2
#> 339         0         2      2
#> 340         1         1      2
#> 341         1         1      2
#> 342         1         0      2
#> 343         1         0      2
#> 344         1         1      2
#> 345         1         2      2
#> 346         1         0      2
#> 347         1         1      2
#> 348         1         2      2
#> 349         1         2      2
#> 350         1         2      2
#> 351         1         0      2
#> 352         0         1      2
#> 353         1         1      2
#> 354         1         0      2
#> 355         1         2      2
#> 356         1         1      2
#> 357         1         0      2
#> 358         1         2      2
#> 359         1         1      2
#> 360         0         0      2
#> 361         1         1      2
#> 362         0         2      2
#> 363         1         0      2
#> 364         1         2      2
#> 365         1         0      2
#> 366         1         1      2
#> 367         1         2      2
#> 368         1         0      2
#> 369         1         2      2
#> 370         0         1      2
#> 371         1         0      2
#> 372         1         1      2
#> 373         1         0      2
#> 374         1         1      2
#> 375         0         2      2
#> 376         0         0      2
#> 377         1         1      2
#> 378         1         2      2
#> 379         0         2      2
#> 380         1         2      2
#> 381         1         0      2
#> 382         0         1      2
#> 383         1         0      2
#> 384         0         1      2
#> 385         1         2      2
#> 386         1         0      2
#> 387         1         1      2
#> 388         1         0      2
#> 389         1         2      2
#> 390         1         1      2
#> 391         1         1      2
#> 392         1         2      2
#> 393         1         0      2
#> 394         0         1      2
#> 395         1         2      2
#> 396         1         0      2
#> 397         0         0      2
#> 398         1         1      2
#> 399         1         2      2
#> 400         1         2      2
#> 401         1         1      2
#> 402         0         0      2
#> 403         0         0      2
#> 404         1         1      2
#> 405         0         1      2
#> 406         0         2      2
#> 407         0         2      2
#> 408         1         0      2
#> 409         0         0      2
#> 410         1         2      2
#> 411         1         1      2
#> 412         1         1      2
#> 413         1         0      2
#> 414         1         2      2
#> 415         1         1      2
#> 416         1         2      2
#> 417         1         2      2
#> 418         0         0      2
#> 419         1         0      2
#> 420         1         1      2
#> 421         1         1      2
#> 422         1         2      2
#> 423         1         0      2
#> 424         1         0      2
#> 425         1         2      2
#> 426         1         1      2
#> 427         0         2      2
#> 428         1         0      2
#> 429         1         0      2
#> 430         1         1      2
#> 431         1         1      2
#> 432         1         2      2
#> 433         1         1      2
#> 434         1         2      2
#> 435         1         1      2
#> 436         1         0      2
#> 437         1         2      2
#> 438         1         0      2
#> 439         1         2      2
#> 440         1         2      2
#> 441         0         1      2
#> 442         0         0      2
#> 443         1         0      2
#> 444         0         1      2
#> 445         0         0      2
#> 446         0         0      2
#> 447         1         1      2
#> 448         1         1      2
#> 449         1         2      2
#> 450         1         2      2
#> 451         1         0      3
#> 452         1         0      3
#> 453         1         2      3
#> 454         1         2      3
#> 455         1         2      3
#> 456         1         0      3
#> 457         1         0      3
#> 458         1         2      3
#> 459         1         0      3
#> 460         1         2      3
#> 461         1         0      3
#> 462         1         2      3
#> 463         1         0      3
#> 464         1         0      3
#> 465         1         2      3
#> 466         1         2      3
#> 467         1         2      3
#> 468         1         2      3
#> 469         1         0      3
#> 470         1         0      3
#> 471         1         2      3
#> 472         1         2      3
#> 473         1         0      3
#> 474         0         0      3
#> 475         1         2      3
#> 476         1         0      3
#> 477         1         0      3
#> 478         0         2      3
#> 479         1         2      3
#> 480         0         0      3
#> 481         1         0      3
#> 482         0         2      3
#> 483         1         2      3
#> 484         1         2      3
#> 485         1         0      3
#> 486         1         0      3
#> 487         1         0      3
#> 488         1         2      3
#> 489         0         0      3
#> 490         1         2      3
#> 491         1         0      3
#> 492         1         2      3
#> 493         1         2      3
#> 494         0         0      3
#> 495         1         2      3
#> 496         1         2      3
#> 497         1         0      3
#> 498         1         0      3
#> 499         1         0      3
#> 500         0         2      3
#> 501         1         0      3
#> 502         1         2      3
#> 503         1         2      3
#> 504         1         2      3
#> 505         1         0      3
#> 506         1         0      3
#> 507         1         0      3
#> 508         0         2      3
#> 509         1         2      3
#> 510         1         0      3
#> 511         1         2      3
#> 512         0         2      3
#> 513         0         0      3
#> 514         1         0      3
#> 515         1         0      3
#> 516         0         2      3
#> 517         1         0      3
#> 518         1         2      3
#> 519         1         0      3
#> 520         1         2      3
#> 521         1         0      3
#> 522         1         2      3
#> 523         1         0      3
#> 524         1         0      3
#> 525         0         2      3
#> 526         1         2      3
#> 527         1         0      3
#> 528         1         2      3
#> 529         1         2      3
#> 530         1         0      3
#> 531         1         2      3
#> 532         1         0      3
#> 533         0         2      3
#> 534         1         0      3
#> 535         1         0      3
#> 536         1         2      3
#> 537         1         0      3
#> 538         1         2      3
#> 539         1         2      3
#> 540         1         2      3
#> 541         1         0      3
#> 542         1         0      3
#> 543         1         2      3
#> 544         1         2      3
#> 545         0         0      3
#> 546         0         0      3
#> 547         1         0      3
#> 548         1         2      3
#> 549         1         0      3
#> 550         1         2      3
#> 551         0         2      4
#> 552         0         3      4
#> 553         1         3      4
#> 554         1         2      4
#> 555         1         0      4
#> 556         0         0      4
#> 557         1         0      4
#> 558         1         3      4
#> 559         0         2      4
#> 560         1         2      4
#> 561         1         3      4
#> 562         1         0      4
#> 563         0         0      4
#> 564         0         0      4
#> 565         1         3      4
#> 566         0         2      4
#> 567         0         3      4
#> 568         1         2      4
#> 569         0         0      4
#> 570         1         3      4
#> 571         1         3      4
#> 572         1         2      4
#> 573         1         2      4
#> 574         1         0      4
#> 575         1         3      4
#> 576         1         3      4
#> 577         1         2      4
#> 578         0         2      4
#> 579         1         0      4
#> 580         1         0      4
#> 581         1         2      4
#> 582         1         2      4
#> 583         0         3      4
#> 584         1         0      4
#> 585         0         3      4
#> 586         1         0      4
#> 587         0         0      4
#> 588         1         3      4
#> 589         0         2      4
#> 590         1         0      4
#> 591         0         3      4
#> 592         0         2      4
#> 593         1         3      4
#> 594         1         0      4
#> 595         1         0      4
#> 596         0         3      4
#> 597         1         2      4
#> 598         1         2      4
#> 599         1         3      4
#> 600         1         0      4
#> 601         1         3      4
#> 602         1         2      4
#> 603         1         0      4
#> 604         1         2      4
#> 605         0         2      4
#> 606         0         0      4
#> 607         1         2      4
#> 608         1         0      4
#> 609         1         3      4
#> 610         0         3      4
#> 611         1         2      4
#> 612         1         2      4
#> 613         0         3      4
#> 614         1         3      4
#> 615         0         0      4
#> 616         1         0      4
#> 617         1         2      4
#> 618         0         3      4
#> 619         0         0      4
#> 620         1         3      4
#> 621         1         0      4
#> 622         1         2      4
#> 623         1         0      4
#> 624         1         2      4
#> 625         1         3      4
#> 626         1         0      4
#> 627         1         2      4
#> 628         1         3      4
#> 629         1         0      4
#> 630         1         2      4
#> 631         1         3      4
#> 632         1         0      4
#> 633         1         2      4
#> 634         1         3      4
#> 635         1         2      4
#> 636         0         0      4
#> 637         1         2      4
#> 638         1         3      4
#> 639         0         0      4
#> 640         1         3      4
#> 641         1         3      4
#> 642         1         2      4
#> 643         1         3      4
#> 644         1         2      4
#> 645         1         0      4
#> 646         1         0      4
#> 647         1         0      4
#> 648         1         2      4
#> 649         0         0      4
#> 650         1         2      4
#> 651         0         3      4
#> 652         1         3      4
#> 653         1         2      4
#> 654         1         2      4
#> 655         0         0      4
#> 656         1         3      4
#> 657         1         3      4
#> 658         1         0      4
#> 659         1         3      4
#> 660         1         2      4
#> 661         1         0      4
#> 662         0         2      4
#> 663         1         0      4
#> 664         0         3      4
#> 665         1         3      4
#> 666         1         2      4
#> 667         1         0      4
#> 668         1         3      4
#> 669         1         0      4
#> 670         1         2      4
#> 671         1         2      4
#> 672         1         0      4
#> 673         1         0      4
#> 674         1         2      4
#> 675         1         3      4
#> 676         1         3      4
#> 677         1         2      4
#> 678         1         2      4
#> 679         0         0      4
#> 680         0         3      4
#> 681         1         0      4
#> 682         1         3      4
#> 683         1         2      4
#> 684         1         3      4
#> 685         1         2      4
#> 686         0         0      4
#> 687         1         0      4
#> 688         1         3      4
#> 689         1         3      4
#> 690         0         0      4
#> 691         1         2      4
#> 692         1         3      4
#> 693         1         0      4
#> 694         0         2      4
#> 695         1         2      4
#> 696         0         2      4
#> 697         0         0      4
#> 698         1         3      4
#> 699         1         3      4
#> 700         1         0      4
#> 701         1         2      4
#> 702         1         3      4
#> 703         0         2      4
#> 704         1         0      4
#> 705         1         0      4
#> 706         0         3      4
#> 707         0         0      4
#> 708         0         3      4
#> 709         1         0      4
#> 710         1         2      4
#> 711         1         2      4
#> 712         1         3      4
#> 713         1         2      4
#> 714         1         0      4
#> 715         1         2      4
#> 716         1         3      4
#> 717         1         3      4
#> 718         1         0      4
#> 719         1         0      4
#> 720         1         3      4
#> 721         1         0      4
#> 722         1         2      4
#> 723         1         3      4
#> 724         0         2      4
#> 725         1         0      4
#> 726         1         3      4
#> 727         1         0      4
#> 728         1         2      4
#> 729         1         3      4
#> 730         0         2      4
#> 731         0         3      4
#> 732         1         0      4
#> 733         1         2      4
#> 734         1         2      4
#> 735         1         3      4
#> 736         1         0      4
#> 737         0         3      4
#> 738         1         2      4
#> 739         1         2      4
#> 740         0         0      4
#> 741         0         0      4
#> 742         1         3      4
#> 743         0         0      4
#> 744         1         3      4
#> 745         1         3      4
#> 746         1         2      4
#> 747         1         0      4
#> 748         1         2      4
#> 749         1         0      4
#> 750         1         2      4
#> 751         1         2      4
#> 752         0         3      4
#> 753         1         3      4
#> 754         1         0      4
#> 755         0         2      4
#> 756         0         0      4
#> 757         1         3      4
#> 758         1         3      4
#> 759         1         0      4
#> 760         0         2      4
#> 761         1         0      5
#> 762         1         3      5
#> 763         0         3      5
#> 764         1         0      5
#> 765         1         0      5
#> 766         1         3      5
#> 767         1         3      5
#> 768         1         0      5
#> 769         0         3      5
#> 770         0         0      5
#> 771         1         3      5
#> 772         1         0      5
#> 773         0         0      5
#> 774         1         3      5
#> 775         1         3      5
#> 776         0         0      5
#> 777         1         0      5
#> 778         1         3      5
#> 779         1         3      5
#> 780         1         0      5
#> 781         1         3      5
#> 782         0         0      5
#> 783         1         0      5
#> 784         1         3      5
#> 785         1         0      5
#> 786         1         3      5
#> 787         1         3      5
#> 788         1         0      5
#> 789         1         0      5
#> 790         1         3      5
#> 791         1         0      5
#> 792         1         3      5
#> 793         0         3      5
#> 794         1         3      5
#> 795         1         0      5
#> 796         1         0      5
#> 797         0         0      5
#> 798         1         3      5
#> 799         1         3      5
#> 800         1         0      5
#> 801         1         0      5
#> 802         1         3      5
#> 803         1         3      5
#> 804         0         0      5
#> 805         1         3      5
#> 806         1         3      5
#> 807         1         0      5
#> 808         1         0      5
#> 809         0         0      5
#> 810         0         3      5
#> 811         1         0      5
#> 812         1         3      5
#> 813         1         0      5
#> 814         1         3      5
#> 815         0         0      5
#> 816         1         3      5
#> 817         1         0      5
#> 818         1         3      5
#> 819         0         0      5
#> 820         1         3      5
#> 821         1         0      5
#> 822         1         3      5
#> 823         1         3      5
#> 824         0         0      5
#> 825         1         0      5
#> 826         1         0      5
#> 827         1         3      5
#> 828         1         3      5
#> 829         1         3      5
#> 830         0         0      5
#> 831         0         0      5
#> 832         1         3      5
#> 833         1         0      5
#> 834         1         3      5
#> 835         1         3      5
#> 836         0         0      5
#> 837         1         0      5
#> 838         0         3      5
#> 839         0         0      5
#> 840         1         3      5
#> 841         1         0      5
#> 842         1         3      5
#> 843         1         3      5
#> 844         0         0      5
#> 845         0         3      5
#> 846         1         3      5
#> 847         1         0      5
#> 848         0         0      5
#> 849         1         3      5
#> 850         1         0      5
#> 851         1         3      5
#> 852         1         0      5
#> 853         1         3      5
#> 854         1         0      5
#> 855         1         0      5
#> 856         1         3      5
#> 857         1         3      5
#> 858         1         0      5
#> 859         1         3      5
#> 860         0         0      5
#> 861         1         3      5
#> 862         1         0      5
#> 863         0         0      5
#> 864         1         3      5
#> 865         1         3      5
#> 866         1         0      5
#> 867         0         3      5
#> 868         1         0      5
#> 869         0         3      5
#> 870         1         3      5
#> 871         1         0      5
#> 872         1         0      5
#> 873         1         0      5
#> 874         1         3      5
#> 875         1         0      5
#> 876         1         3      5
#> 877         1         0      5
#> 878         0         0      5
#> 879         1         3      5
#> 880         1         3      5
#> 881         1         0      5
#> 882         0         3      5
#> 883         1         3      5
#> 884         1         0      5
#> 885         1         3      5
#> 886         0         0      5
#> 887         1         0      5
#> 888         0         3      5
#> 889         0         3      5
#> 890         1         0      5
#> 891         1         3      5
#> 892         0         0      5
#> 893         0         3      5
#> 894         1         0      5
#> 895         1         0      5
#> 896         1         3      5
#> 897         1         0      5
#> 898         1         3      5
#> 899         0         3      5
#> 900         0         0      5
#> 901         1         3      5
#> 902         1         0      5
#> 903         1         0      5
#> 904         1         3      5
#> 905         1         3      5
#> 906         1         0      5
#> 907         1         3      5
#> 908         1         0      5
#> 909         1         0      5
#> 910         1         3      5
#> 911         1         0      5
#> 912         0         3      5
#> 913         1         0      5
#> 914         1         3      5
#> 915         1         0      5
#> 916         1         3      5
#> 917         0         0      5
#> 918         1         0      5
#> 919         1         3      5
#> 920         1         3      5
#> 921         1         3      5
#> 922         0         0      5
#> 923         1         0      5
#> 924         0         3      5
#> 925         1         3      5
#> 926         1         3      5
#> 927         1         0      5
#> 928         1         0      5
#> 929         1         0      5
#> 930         1         3      5
#> 931         0         0      5
#> 932         1         3      5
#> 933         1         3      5
#> 934         1         3      5
#> 935         1         0      5
#> 936         1         0      5
#> 937         1         0      5
#> 938         1         3      5
#> 939         1         0      5
#> 940         1         3      5
#> 941         1         0      5
#> 942         1         0      5
#> 943         1         3      5
#> 944         0         3      5
#> 945         1         3      5
#> 946         0         0      5
#> 947         1         3      5
#> 948         1         0      5
#> 949         1         0      5
#> 950         1         3      5
#> 951         0         0      5
#> 952         1         3      5
#> 953         1         3      5
#> 954         0         0      5
#> 955         1         3      5
#> 956         0         0      5
#> 957         1         0      5
#> 958         0         3      5
#> 959         1         0      5
#> 960         1         3      5
#> 961         1         3      5
#> 962         1         3      5
#> 963         1         0      5
#> 964         1         0      5
#> 965         1         3      5
#> 966         0         3      5
#> 967         0         0      5
#> 968         1         0      5
#> 969         1         0      5
#> 970         0         0      5
#> 971         0         3      5
#> 972         0         3      5
#> 973         1         3      5
#> 974         1         0      5
#> 975         0         3      5
#> 976         1         0      5
#> 977         0         3      5
#> 978         0         0      5
#> 979         1         0      5
#> 980         1         3      5
#> 981         1         0      5
#> 982         1         3      5
#> 983         1         0      5
#> 984         1         3      5
#> 985         1         0      5
#> 986         1         3      5
#> 987         1         0      5
#> 988         1         3      5
#> 989         1         0      5
#> 990         0         0      5
#> 991         0         3      5
#> 992         1         3      5
#> 993         0         3      5
#> 994         1         3      5
#> 995         0         0      5
#> 996         1         0      5
#> 997         0         0      5
#> 998         0         3      5
#> 999         0         0      5
#> 1000        0         3      5

# Usage option 2 - specify n_arm and alloc_ratios
head(datasim_bin(n_total = NULL, num_arms = NULL, d = NULL, n_arm = 200, alloc_ratios = matrix(c(1,1,1,1,1,NA,NA,1,1), ncol = 3, byrow = T), period_blocks = 2, p0 = 0.7, OR = rep(1.4, 3), lambda = rep(0.15, 4), trend = "linear"))
#>   response treatment period
#> 1        1         0      1
#> 2        1         1      1
#> 3        1         1      1
#> 4        1         0      1
#> 5        1         0      1
#> 6        1         0      1