Frequentist linear regression model analysis for continuous data adjusting for calendar time units

This function performs linear regression taking into account all trial data until the arm under study leaves the trial and adjusting for calendar time units as factors.

Usage

fixmodel_cal_cont(
  data,
  arm,
  alpha = 0.025,
  unit_size = 25,
  ncc = TRUE,
  check = TRUE,
  ...
)

Arguments

data

Data frame with trial data, e.g. result from the datasim_cont() function. Must contain columns named 'treatment' and 'response'.

arm

Integer. Index of the treatment arm under study to perform inference on (vector of length 1). This arm is compared to the control group.

alpha

Double. Significance level (one-sided). Default=0.025.

unit_size

Integer. Number of patients per calendar time unit. Default=25.

ncc

Logical. Indicates whether to include non-concurrent data into the analysis. Default=TRUE.

check

Logical. Indicates whether the input parameters should be checked by the function. Default=TRUE, unless the function is called by a simulation function, where the default is FALSE.

...

Further arguments passed by wrapper functions when running simulations.

Value

List containing the following elements regarding the results of comparing arm to control:

p-val - p-value (one-sided)

treat_effect - estimated treatment effect in terms of the difference in means

lower_ci - lower limit of the (1-2*alpha)*100% confidence interval

upper_ci - upper limit of the (1-2*alpha)*100% confidence interval

reject_h0 - indicator of whether the null hypothesis was rejected or not (p_val < alpha)

model - fitted model

Details

The model-based analysis adjusts for the time effect by including the factor calendar time unit (defined as time units of fixed length, defined by ùnit_size). The time is then modelled as a step-function with jumps at the beginning of each calendar time unit. Denoting by \(y_j\) the continuous response for patient \(j\), by \(k_j\) the arm patient \(j\) was allocated to, and by \(M\) the treatment arm under evaluation, the regression model is given by:

$$E(y_j) = \eta_0 + \sum_{k \in \mathcal{K}_M} \theta_k \cdot I(k_j=k) + \sum_{c=2}^{C_M} \tau_c \cdot I(t_j \in T_{C_c})$$

where \(\eta_0\) is the response in the control arm in the first calendar time unit; \(\theta_k\) represents the effect of treatment \(k\) compared to control for \(k\in\mathcal{K}_M\), where \(\mathcal{K}_M\) is the set of treatments that were active in the trial during calendar time units prior or up to the time when the investigated treatment arm left the trial; \(\tau_c\) indicates the stepwise calendar time effect between calendar time units 1 and \(c\) (\(c = 2, \ldots, C_M\)), where \(C_M\) denotes the calendar time unit, in which the investigated treatment arm left the trial.

If the data consists of only one calendar time unit, the calendar time unit in not used as covariate.

Author

Pavla Krotka

Examples


trial_data <- datasim_cont(num_arms = 3, n_arm = 100, d = c(0, 100, 250),
theta = rep(0.25, 3), lambda = rep(0.15, 4), sigma = 1, trend = "linear")

fixmodel_cal_cont(data = trial_data, arm = 3)
#> $p_val
#> [1] 0.0007310476
#> 
#> $treat_effect
#> [1] 0.4112412
#> 
#> $lower_ci
#> [1] 0.1587832
#> 
#> $upper_ci
#> [1] 0.6636993
#> 
#> $reject_h0
#> [1] TRUE
#> 
#> $model
#> 
#> Call:
#> lm(formula = response ~ as.factor(treatment) + as.factor(cal_time), 
#>     data = data_new)
#> 
#> Coefficients:
#>           (Intercept)  as.factor(treatment)1  as.factor(treatment)2  
#>               0.25594                0.23671                0.27302  
#> as.factor(treatment)3   as.factor(cal_time)2   as.factor(cal_time)3  
#>               0.41124               -0.38591                0.02469  
#>  as.factor(cal_time)4   as.factor(cal_time)5   as.factor(cal_time)6  
#>              -0.04598               -0.26118               -0.08639  
#>  as.factor(cal_time)7   as.factor(cal_time)8   as.factor(cal_time)9  
#>              -0.05793               -0.03381               -0.08750  
#> as.factor(cal_time)10  as.factor(cal_time)11  as.factor(cal_time)12  
#>              -0.18016               -0.61170                0.04718  
#> as.factor(cal_time)13  as.factor(cal_time)14  as.factor(cal_time)15  
#>               0.05044               -0.47542                0.15559  
#> as.factor(cal_time)16  as.factor(cal_time)17  as.factor(cal_time)18  
#>               0.03035               -0.10822                0.06574  
#> as.factor(cal_time)19  as.factor(cal_time)20  
#>              -0.40871               -0.33677  
#> 
#>