Model-based analysis for continuous data using discontinuous piecewise polynomials per calendar time unit

This function performs linear regression taking into account all trial data until the arm under study leaves the trial and adjusting for time using discontinuous piecewise polynomials in each calendar time unit.

Usage

piecewise_cal_cont(
  data,
  arm,
  alpha = 0.025,
  unit_size = 25,
  ncc = TRUE,
  poly_degree = 3,
  check = TRUE,
  ...
)

Arguments

data

Data frame with trial data, e.g. result from the datasim_cont() function. Must contain columns named 'treatment', 'response' and 'j'.

arm

Integer. Index of the treatment arm under study to perform inference on (vector of length 1). This arm is compared to the control group.

alpha

Double. Significance level (one-sided). Default=0.025.

unit_size

Integer. Number of patients per calendar time unit. Default=25.

ncc

Logical. Indicates whether to include non-concurrent data into the analysis. Default=TRUE.

poly_degree

Integer. Degree of the piecewise polynomial. Default=3.

check

Logical. Indicates whether the input parameters should be checked by the function. Default=TRUE, unless the function is called by a simulation function, where the default is FALSE.

...

Further arguments passed by wrapper functions when running simulations.

Value

List containing the following elements regarding the results of comparing arm to control:

p-val - p-value (one-sided)

treat_effect - estimated treatment effect in terms of the difference in means

lower_ci - lower limit of the (1-2*alpha)*100% confidence interval

upper_ci - upper limit of the (1-2*alpha)*100% confidence interval

reject_h0 - indicator of whether the null hypothesis was rejected or not (p_val < alpha)

model - fitted model

Author

Pavla Krotka

Examples


trial_data <- datasim_cont(num_arms = 3, n_arm = 100, d = c(0, 100, 250),
theta = rep(0.25, 3), lambda = rep(0.15, 4), sigma = 1, trend = "linear")

piecewise_cal_cont(data = trial_data, arm = 3)
#> $p_val
#> [1] 0.1992429
#> 
#> $treat_effect
#> [1] 0.1227509
#> 
#> $lower_ci
#> [1] -0.1627302
#> 
#> $upper_ci
#> [1] 0.408232
#> 
#> $reject_h0
#> [1] FALSE
#> 
#> $model
#> 
#> Call:
#> lm(formula = response ~ as.factor(treatment) + poly(j, degree = poly_degree, 
#>     raw = TRUE) * as.factor(cal_time), data = data_new)
#> 
#> Coefficients:
#>                                                      (Intercept)  
#>                                                        3.381e-01  
#>                                            as.factor(treatment)1  
#>                                                        3.731e-01  
#>                                            as.factor(treatment)2  
#>                                                        6.503e-02  
#>                                            as.factor(treatment)3  
#>                                                        1.228e-01  
#>                       poly(j, degree = poly_degree, raw = TRUE)1  
#>                                                       -4.994e-02  
#>                       poly(j, degree = poly_degree, raw = TRUE)2  
#>                                                        1.188e-03  
#>                       poly(j, degree = poly_degree, raw = TRUE)3  
#>                                                        8.708e-05  
#>                                             as.factor(cal_time)2  
#>                                                        4.987e+00  
#>                                             as.factor(cal_time)3  
#>                                                        2.463e+02  
#>                                             as.factor(cal_time)4  
#>                                                       -7.746e+01  
#>                                             as.factor(cal_time)5  
#>                                                       -6.474e+02  
#>                                             as.factor(cal_time)6  
#>                                                        3.584e+01  
#>                                             as.factor(cal_time)7  
#>                                                        2.391e+03  
#>                                             as.factor(cal_time)8  
#>                                                       -1.999e+03  
#>                                             as.factor(cal_time)9  
#>                                                        6.759e+02  
#>                                            as.factor(cal_time)10  
#>                                                        5.904e+03  
#>                                            as.factor(cal_time)11  
#>                                                       -2.428e+03  
#>                                            as.factor(cal_time)12  
#>                                                       -1.653e+04  
#>                                            as.factor(cal_time)13  
#>                                                       -1.528e+03  
#>                                            as.factor(cal_time)14  
#>                                                        1.017e+03  
#>                                            as.factor(cal_time)15  
#>                                                       -3.244e+04  
#>                                            as.factor(cal_time)16  
#>                                                        2.354e+04  
#>                                            as.factor(cal_time)17  
#>                                                       -1.713e+04  
#>                                            as.factor(cal_time)18  
#>                                                       -1.138e+04  
#>                                            as.factor(cal_time)19  
#>                                                       -6.106e+04  
#>                                            as.factor(cal_time)20  
#>                                                        6.329e+04  
#>  poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)2  
#>                                                       -5.292e-01  
#>  poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)2  
#>                                                        1.670e-02  
#>  poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)2  
#>                                                       -2.499e-04  
#>  poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)3  
#>                                                       -1.173e+01  
#>  poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)3  
#>                                                        1.848e-01  
#>  poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)3  
#>                                                       -1.059e-03  
#>  poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)4  
#>                                                        2.807e+00  
#>  poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)4  
#>                                                       -3.391e-02  
#>  poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)4  
#>                                                        4.121e-05  
#>  poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)5  
#>                                                        1.689e+01  
#>  poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)5  
#>                                                       -1.467e-01  
#>  poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)5  
#>                                                        3.310e-04  
#>  poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)6  
#>                                                        1.802e-01  
#>  poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)6  
#>                                                       -8.847e-03  
#>  poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)6  
#>                                                       -5.236e-05  
#>  poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)7  
#>                                                       -4.478e+01  
#>  poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)7  
#>                                                        2.783e-01  
#>  poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)7  
#>                                                       -6.663e-04  
#>  poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)8  
#>                                                        3.192e+01  
#>  poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)8  
#>                                                       -1.708e-01  
#>  poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)8  
#>                                                        2.141e-04  
#>  poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)9  
#>                                                       -8.924e+00  
#>  poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)9  
#>                                                        3.829e-02  
#>  poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)9  
#>                                                       -1.446e-04  
#> poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)10  
#>                                                       -7.637e+01  
#> poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)10  
#>                                                        3.282e-01  
#> poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)10  
#>                                                       -5.598e-04  
#> poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)11  
#>                                                        2.666e+01  
#> poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)11  
#>                                                       -9.817e-02  
#> poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)11  
#>                                                        3.047e-05  
#> poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)12  
#>                                                        1.727e+02  
#> poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)12  
#>                                                       -6.021e-01  
#> poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)12  
#>                                                        6.094e-04  
#> poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)13  
#>                                                        1.564e+01  
#> poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)13  
#>                                                       -5.409e-02  
#> poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)13  
#>                                                       -2.740e-05  
#> poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)14  
#>                                                       -1.065e+01  
#> poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)14  
#>                                                        3.529e-02  
#> poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)14  
#>                                                       -1.277e-04  
#> poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)15  
#>                                                        2.672e+02  
#> poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)15  
#>                                                       -7.344e-01  
#> poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)15  
#>                                                        5.836e-04  
#> poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)16  
#>                                                       -1.821e+02  
#> poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)16  
#>                                                        4.688e-01  
#> poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)16  
#>                                                       -4.911e-04  
#> poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)17  
#>                                                        1.225e+02  
#> poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)17  
#>                                                       -2.930e-01  
#> poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)17  
#>                                                        1.447e-04  
#> poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)18  
#>                                                        7.667e+01  
#> poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)18  
#>                                                       -1.730e-01  
#> poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)18  
#>                                                        4.124e-05  
#> poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)19  
#>                                                        3.956e+02  
#> poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)19  
#>                                                       -8.551e-01  
#> poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)19  
#>                                                        5.273e-04  
#> poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)20  
#>                                                       -3.894e+02  
#> poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)20  
#>                                                        7.975e-01  
#> poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)20  
#>                                                       -6.330e-04  
#> 
#>