Pooled analysis for binary data

This function performs pooled analysis (naively pooling concurrent and non-concurrent controls without adjustment) using a logistic model.

Usage

poolmodel_bin(data, arm, alpha = 0.025, check = TRUE, ...)

Arguments

data

Trial data, e.g. result from the datasim_bin() function. Must contain columns named 'treatment', 'response' and 'period'.

arm

Indicator of the treatment arm under study to perform inference on (vector of length 1). This arm is compared to the control group.

alpha

Significance level. Default=0.025.

check

Boolean. Indicates whether the input parameters should be checked by the function. Default=TRUE, unless the function is called by a simulation function, where the default is FALSE.

...

Further arguments for simulation function.

Value

List containing the following elements regarding the results of comparing arm to control:

p-val - p-value (one-sided)

treat_effect - estimated treatment effect in terms of the log-odds ratio

lower_ci - lower limit of the 95% confidence interval

upper_ci - upper limit of the 95% confidence interval

reject_h0 - indicator of whether the null hypothesis was rejected or not (p_val < alpha)

model - fitted model

Author

Pavla Krotka

Examples


trial_data <- datasim_bin(num_arms = 3, n_arm = 100, d = c(0, 100, 250),
p0 = 0.7, OR = rep(1.8, 3), lambda = rep(0.15, 4), trend="stepwise")

poolmodel_bin(data = trial_data, arm = 3)
#> $p_val
#> [1] 0.00276026
#> 
#> $treat_effect
#> [1] 0.8873032
#> 
#> $lower_ci
#> [1] 0.2827163
#> 
#> $upper_ci
#> [1] 1.543606
#> 
#> $reject_h0
#> [1] TRUE
#> 
#> $model
#> 
#> Call:  glm(formula = response ~ as.factor(treatment), family = binomial, 
#>     data = data_new)
#> 
#> Coefficients:
#>           (Intercept)  as.factor(treatment)3  
#>                0.8473                 0.8873  
#> 
#> Degrees of Freedom: 299 Total (i.e. Null);  298 Residual
#> Null Deviance:	    337.4 
#> Residual Deviance: 328.9 	AIC: 332.9
#>