Generation of a seasonal trend

This function generates a time trend for given time points in the trial according to a periodic function.

Usage

seasonal_trend(j, lambda, n_wave, n_total)

Arguments

j

Time points for which the trend should be generated.

lambda

Strength of time trend.

n_wave

How many cycles (waves) should the time trend have (\(\psi\)).

n_total

Total sample size in the trial.

Value

Time trend for time points j.

Details

The time trend is generated according to the function \(f(j) = \lambda \cdot \mathrm{sin} \big( \psi \cdot 2\pi \cdot \frac{j-1}{N-1} \big)\), where the parameter \(\psi\) corresponds to the input parameter n_wave.

Author

Marta Bofill Roig, Pavla Krotka