Spline regression analysis for continuous data with knots placed according to periods

This function performs linear regression taking into account all trial data until the arm under study leaves the trial and adjusting for time using regression splines with knots placed according to periods.

Usage

splines_cont(
  data,
  arm,
  alpha = 0.025,
  ncc = TRUE,
  bs_degree = 3,
  check = TRUE,
  ...
)

Arguments

data

Data frame with trial data, e.g. result from the datasim_cont() function. Must contain columns named 'treatment', 'response', 'period' and 'j'.

arm

Integer. Index of the treatment arm under study to perform inference on (vector of length 1). This arm is compared to the control group.

alpha

Double. Significance level (one-sided). Default=0.025.

ncc

Logical. Indicates whether to include non-concurrent data into the analysis. Default=TRUE.

bs_degree

Integer. Degree of the polynomial spline. Default=3 for cubic spline.

check

Logical. Indicates whether the input parameters should be checked by the function. Default=TRUE, unless the function is called by a simulation function, where the default is FALSE.

...

Further arguments passed by wrapper functions when running simulations.

Value

List containing the following elements regarding the results of comparing arm to control:

p-val - p-value (one-sided)

treat_effect - estimated treatment effect in terms of the difference in means

lower_ci - lower limit of the (1-2*alpha)*100% confidence interval

upper_ci - upper limit of the (1-2*alpha)*100% confidence interval

reject_h0 - indicator of whether the null hypothesis was rejected or not (p_val < alpha)

knots - positions of the knots in terms of patient index

model - fitted model

Author

Pavla Krotka

Examples


trial_data <- datasim_cont(num_arms = 3, n_arm = 100, d = c(0, 100, 250),
theta = rep(0.25, 3), lambda = rep(0.15, 4), sigma = 1, trend = "linear")

splines_cont(data = trial_data, arm = 3)
#> $p_val
#> [1] 0.001838685
#> 
#> $treat_effect
#> [1] 0.4118378
#> 
#> $lower_ci
#> [1] 0.1345897
#> 
#> $upper_ci
#> [1] 0.6890859
#> 
#> $reject_h0
#> [1] TRUE
#> 
#> $knots
#> [1] 100 250 400
#> 
#> $model
#> 
#> Call:
#> lm(formula = response ~ as.factor(treatment) + bs(j, knots = bs_knots, 
#>     degree = bs_degree), data = data_new)
#> 
#> Coefficients:
#>                                  (Intercept)  
#>                                    -0.024590  
#>                        as.factor(treatment)1  
#>                                     0.385147  
#>                        as.factor(treatment)2  
#>                                     0.423823  
#>                        as.factor(treatment)3  
#>                                     0.411838  
#> bs(j, knots = bs_knots, degree = bs_degree)1  
#>                                    -0.005839  
#> bs(j, knots = bs_knots, degree = bs_degree)2  
#>                                     0.085050  
#> bs(j, knots = bs_knots, degree = bs_degree)3  
#>                                    -0.099641  
#> bs(j, knots = bs_knots, degree = bs_degree)4  
#>                                     0.081261  
#> bs(j, knots = bs_knots, degree = bs_degree)5  
#>                                    -0.254250  
#> bs(j, knots = bs_knots, degree = bs_degree)6  
#>                                    -0.041273  
#> 
#>