Data simulation for binary endpoints for a platform trial with an arbitrary number of treatment arms entering sequentially

Simulates data from a platform trial with binary endpoints and an arbitrary number of treatment arms entering sequentially. There is the option to either specify the timing of adding arms and the overall sample size in the trial or specify the sample size per arm (assumed equal) and the allocation ratios per period.

datasim_bin(
  n_total,
  num_arms,
  d,
  n_arm,
  alloc_ratios,
  period_blocks = 2,
  p0,
  OR,
  lambda,
  trend,
  N_peak,
  full = FALSE
)

Arguments

n_total

Overall sample size for the trial

num_arms

Number of treatment arms in the trial

d

Timing of adding new arms in terms of number of patients allocated to the control arm

n_arm

Sample size per arm

alloc_ratios

Matrix with allocation ratios in each period (rows = arms; columns = periods)

period_blocks

Number to multiply the number of active arms with in order to get the block size per period

p0

Response in the control arm

OR

Vector with odds ratios for each arm

lambda

Vector with strength of time trend in each arm

trend

Indicates the time trend pattern ("linear", "stepwise" or "inv_u")

N_peak

Point at which the inverted-u time trend switches direction in terms of overall sample size

full

Boolean. Indicates whether only variables needed for the analysis should be in the output (FALSE) or also additional information (lambdas, underlying responses) should be included (TRUE). Default=FALSE

Value

Data frame: simulated trial data

Author

Pavla Krotka, Marta Bofill Roig

Examples


# Usage option 1 - specify n_total, num_arms and d
datasim_bin(n_total = 1000, num_arms = 3, d = 120,
n_arm = NULL, alloc_ratios = NULL,
period_blocks = 2, p0 = 0.7, OR = rep(1.4, 3), lambda = rep(0.15, 4), trend = "linear")
#> Loading required package: rlang
#> Warning: package 'rlang' was built under R version 3.6.3
#>      response treatment period
#> 1           1         0      1
#> 2           1         0      1
#> 3           0         1      1
#> 4           1         1      1
#> 5           0         1      1
#> 6           1         0      1
#> 7           1         0      1
#> 8           0         1      1
#> 9           1         0      1
#> 10          1         1      1
#> 11          1         1      1
#> 12          0         0      1
#> 13          0         1      1
#> 14          1         1      1
#> 15          1         0      1
#> 16          1         0      1
#> 17          1         0      1
#> 18          1         1      1
#> 19          0         0      1
#> 20          1         1      1
#> 21          1         1      1
#> 22          0         0      1
#> 23          1         1      1
#> 24          1         0      1
#> 25          1         0      1
#> 26          0         1      1
#> 27          0         0      1
#> 28          1         1      1
#> 29          1         0      1
#> 30          0         1      1
#> 31          0         0      1
#> 32          0         1      1
#> 33          0         1      1
#> 34          1         0      1
#> 35          1         0      1
#> 36          0         1      1
#> 37          1         1      1
#> 38          0         0      1
#> 39          1         1      1
#> 40          1         0      1
#> 41          1         0      1
#> 42          1         0      1
#> 43          1         1      1
#> 44          1         1      1
#> 45          1         1      1
#> 46          0         1      1
#> 47          1         0      1
#> 48          1         0      1
#> 49          1         0      1
#> 50          1         1      1
#> 51          1         1      1
#> 52          1         0      1
#> 53          1         0      1
#> 54          0         1      1
#> 55          1         1      1
#> 56          1         0      1
#> 57          1         0      1
#> 58          1         1      1
#> 59          1         0      1
#> 60          0         1      1
#> 61          0         0      1
#> 62          0         1      1
#> 63          0         1      1
#> 64          1         0      1
#> 65          1         1      1
#> 66          0         0      1
#> 67          0         0      1
#> 68          1         1      1
#> 69          1         1      1
#> 70          1         0      1
#> 71          1         1      1
#> 72          1         0      1
#> 73          1         0      1
#> 74          1         0      1
#> 75          1         1      1
#> 76          1         1      1
#> 77          1         0      1
#> 78          1         1      1
#> 79          1         0      1
#> 80          1         1      1
#> 81          1         1      1
#> 82          0         0      1
#> 83          1         1      1
#> 84          1         0      1
#> 85          1         0      1
#> 86          1         0      1
#> 87          1         1      1
#> 88          1         1      1
#> 89          1         1      1
#> 90          0         0      1
#> 91          1         1      1
#> 92          0         0      1
#> 93          1         1      1
#> 94          1         1      1
#> 95          0         0      1
#> 96          1         0      1
#> 97          1         0      1
#> 98          0         0      1
#> 99          1         1      1
#> 100         0         1      1
#> 101         1         0      1
#> 102         1         1      1
#> 103         1         0      1
#> 104         1         1      1
#> 105         0         1      1
#> 106         1         1      1
#> 107         1         0      1
#> 108         1         0      1
#> 109         1         1      1
#> 110         1         0      1
#> 111         1         0      1
#> 112         1         1      1
#> 113         1         1      1
#> 114         0         0      1
#> 115         1         1      1
#> 116         0         0      1
#> 117         1         1      1
#> 118         1         0      1
#> 119         0         1      1
#> 120         1         0      1
#> 121         0         0      1
#> 122         0         1      1
#> 123         1         0      1
#> 124         1         1      1
#> 125         1         0      1
#> 126         0         1      1
#> 127         0         1      1
#> 128         1         0      1
#> 129         1         1      1
#> 130         1         0      1
#> 131         1         1      1
#> 132         0         0      1
#> 133         1         0      1
#> 134         1         0      1
#> 135         0         1      1
#> 136         1         1      1
#> 137         1         0      1
#> 138         1         1      1
#> 139         1         0      1
#> 140         1         1      1
#> 141         1         0      1
#> 142         1         0      1
#> 143         1         1      1
#> 144         0         1      1
#> 145         1         0      1
#> 146         1         1      1
#> 147         1         0      1
#> 148         1         1      1
#> 149         1         1      1
#> 150         1         0      1
#> 151         0         1      1
#> 152         0         0      1
#> 153         0         1      1
#> 154         1         1      1
#> 155         1         0      1
#> 156         1         0      1
#> 157         1         0      1
#> 158         0         0      1
#> 159         0         1      1
#> 160         1         1      1
#> 161         0         1      1
#> 162         1         1      1
#> 163         1         0      1
#> 164         1         0      1
#> 165         1         1      1
#> 166         1         0      1
#> 167         0         0      1
#> 168         1         1      1
#> 169         1         0      1
#> 170         1         1      1
#> 171         0         0      1
#> 172         1         1      1
#> 173         1         1      1
#> 174         1         1      1
#> 175         1         0      1
#> 176         1         0      1
#> 177         0         0      1
#> 178         1         1      1
#> 179         0         1      1
#> 180         1         0      1
#> 181         1         0      1
#> 182         0         1      1
#> 183         1         1      1
#> 184         1         0      1
#> 185         0         0      1
#> 186         0         1      1
#> 187         1         0      1
#> 188         1         1      1
#> 189         0         0      1
#> 190         1         0      1
#> 191         1         1      1
#> 192         1         1      1
#> 193         1         1      1
#> 194         0         1      1
#> 195         1         0      1
#> 196         1         0      1
#> 197         1         0      1
#> 198         1         0      1
#> 199         1         1      1
#> 200         0         1      1
#> 201         0         0      1
#> 202         1         0      1
#> 203         1         1      1
#> 204         0         1      1
#> 205         1         0      1
#> 206         1         1      1
#> 207         1         1      1
#> 208         0         0      1
#> 209         1         0      1
#> 210         0         0      1
#> 211         1         1      1
#> 212         0         1      1
#> 213         1         0      1
#> 214         1         0      1
#> 215         1         1      1
#> 216         1         1      1
#> 217         0         1      1
#> 218         0         0      1
#> 219         1         1      1
#> 220         0         0      1
#> 221         1         0      1
#> 222         1         1      1
#> 223         1         1      1
#> 224         0         0      1
#> 225         1         0      1
#> 226         0         1      1
#> 227         0         0      1
#> 228         0         1      1
#> 229         1         1      1
#> 230         0         0      1
#> 231         1         1      1
#> 232         1         0      1
#> 233         1         0      1
#> 234         1         1      1
#> 235         1         0      1
#> 236         1         1      1
#> 237         1         1      1
#> 238         1         0      1
#> 239         1         1      1
#> 240         1         0      1
#> 241         0         2      2
#> 242         1         1      2
#> 243         0         0      2
#> 244         1         1      2
#> 245         1         2      2
#> 246         1         0      2
#> 247         1         1      2
#> 248         0         0      2
#> 249         1         1      2
#> 250         1         2      2
#> 251         1         0      2
#> 252         1         2      2
#> 253         1         2      2
#> 254         1         2      2
#> 255         1         0      2
#> 256         1         0      2
#> 257         0         1      2
#> 258         0         1      2
#> 259         1         2      2
#> 260         1         2      2
#> 261         1         0      2
#> 262         0         1      2
#> 263         1         0      2
#> 264         1         1      2
#> 265         0         2      2
#> 266         1         1      2
#> 267         1         1      2
#> 268         1         0      2
#> 269         1         0      2
#> 270         1         2      2
#> 271         1         1      2
#> 272         0         0      2
#> 273         1         2      2
#> 274         1         1      2
#> 275         1         2      2
#> 276         1         0      2
#> 277         1         2      2
#> 278         0         1      2
#> 279         1         0      2
#> 280         0         0      2
#> 281         0         2      2
#> 282         1         1      2
#> 283         1         1      2
#> 284         1         0      2
#> 285         1         2      2
#> 286         0         1      2
#> 287         1         2      2
#> 288         1         0      2
#> 289         1         0      2
#> 290         0         0      2
#> 291         0         2      2
#> 292         1         1      2
#> 293         1         2      2
#> 294         1         1      2
#> 295         0         2      2
#> 296         0         1      2
#> 297         1         0      2
#> 298         0         2      2
#> 299         1         0      2
#> 300         1         1      2
#> 301         1         0      2
#> 302         1         1      2
#> 303         1         1      2
#> 304         1         2      2
#> 305         1         0      2
#> 306         1         2      2
#> 307         1         2      2
#> 308         1         1      2
#> 309         1         0      2
#> 310         1         2      2
#> 311         1         1      2
#> 312         1         0      2
#> 313         1         1      2
#> 314         1         2      2
#> 315         1         2      2
#> 316         1         1      2
#> 317         1         0      2
#> 318         1         0      2
#> 319         1         2      2
#> 320         1         2      2
#> 321         0         1      2
#> 322         1         0      2
#> 323         1         1      2
#> 324         1         0      2
#> 325         0         2      2
#> 326         1         2      2
#> 327         1         1      2
#> 328         1         0      2
#> 329         1         1      2
#> 330         1         0      2
#> 331         1         1      2
#> 332         1         1      2
#> 333         1         2      2
#> 334         0         0      2
#> 335         1         0      2
#> 336         1         2      2
#> 337         1         0      2
#> 338         1         2      2
#> 339         1         0      2
#> 340         1         1      2
#> 341         0         1      2
#> 342         1         2      2
#> 343         0         0      2
#> 344         1         2      2
#> 345         1         1      2
#> 346         0         0      2
#> 347         1         1      2
#> 348         1         2      2
#> 349         1         0      2
#> 350         0         0      2
#> 351         1         2      2
#> 352         1         2      2
#> 353         1         1      2
#> 354         0         1      2
#> 355         0         0      2
#> 356         1         1      2
#> 357         1         2      2
#> 358         1         2      2
#> 359         1         1      2
#> 360         1         0      2
#> 361         0         1      2
#> 362         1         1      2
#> 363         1         0      2
#> 364         1         0      2
#> 365         1         2      2
#> 366         1         2      2
#> 367         0         1      2
#> 368         1         0      2
#> 369         1         0      2
#> 370         1         2      2
#> 371         1         1      2
#> 372         1         2      2
#> 373         0         0      2
#> 374         1         1      2
#> 375         1         0      2
#> 376         0         1      2
#> 377         1         2      2
#> 378         0         2      2
#> 379         1         0      2
#> 380         0         2      2
#> 381         1         1      2
#> 382         1         1      2
#> 383         1         2      2
#> 384         1         0      2
#> 385         1         1      2
#> 386         0         0      2
#> 387         1         0      2
#> 388         1         2      2
#> 389         1         2      2
#> 390         0         1      2
#> 391         1         1      2
#> 392         1         2      2
#> 393         0         1      2
#> 394         0         0      2
#> 395         1         2      2
#> 396         0         0      2
#> 397         0         0      2
#> 398         1         1      2
#> 399         0         0      2
#> 400         1         2      2
#> 401         1         2      2
#> 402         1         1      2
#> 403         1         2      2
#> 404         1         2      2
#> 405         1         1      2
#> 406         1         0      2
#> 407         1         0      2
#> 408         1         1      2
#> 409         1         2      2
#> 410         0         0      2
#> 411         1         1      2
#> 412         1         2      2
#> 413         0         0      2
#> 414         1         1      2
#> 415         1         0      2
#> 416         1         2      2
#> 417         1         2      2
#> 418         1         1      2
#> 419         0         1      2
#> 420         0         0      2
#> 421         1         1      2
#> 422         1         1      2
#> 423         1         2      2
#> 424         1         0      2
#> 425         1         2      2
#> 426         1         0      2
#> 427         1         1      2
#> 428         1         1      2
#> 429         1         2      2
#> 430         1         0      2
#> 431         1         2      2
#> 432         0         0      2
#> 433         1         0      2
#> 434         1         1      2
#> 435         1         2      2
#> 436         1         0      2
#> 437         1         1      2
#> 438         1         2      2
#> 439         0         0      2
#> 440         1         2      2
#> 441         0         2      2
#> 442         1         0      2
#> 443         1         1      2
#> 444         0         1      2
#> 445         0         0      2
#> 446         0         1      2
#> 447         0         1      2
#> 448         1         2      2
#> 449         1         0      2
#> 450         1         2      2
#> 451         1         2      3
#> 452         1         0      3
#> 453         1         2      3
#> 454         1         0      3
#> 455         0         0      3
#> 456         0         2      3
#> 457         1         0      3
#> 458         1         2      3
#> 459         1         0      3
#> 460         1         2      3
#> 461         0         0      3
#> 462         1         2      3
#> 463         1         2      3
#> 464         1         2      3
#> 465         1         0      3
#> 466         1         0      3
#> 467         1         0      3
#> 468         1         0      3
#> 469         1         2      3
#> 470         1         2      3
#> 471         0         2      3
#> 472         1         2      3
#> 473         1         0      3
#> 474         1         0      3
#> 475         1         0      3
#> 476         0         0      3
#> 477         1         2      3
#> 478         0         2      3
#> 479         0         0      3
#> 480         1         0      3
#> 481         1         2      3
#> 482         1         2      3
#> 483         1         0      3
#> 484         1         2      3
#> 485         1         0      3
#> 486         1         2      3
#> 487         1         2      3
#> 488         1         2      3
#> 489         1         0      3
#> 490         1         0      3
#> 491         1         0      3
#> 492         1         2      3
#> 493         0         0      3
#> 494         1         2      3
#> 495         1         2      3
#> 496         1         0      3
#> 497         1         0      3
#> 498         1         2      3
#> 499         1         0      3
#> 500         1         2      3
#> 501         1         0      3
#> 502         0         2      3
#> 503         0         0      3
#> 504         1         2      3
#> 505         1         0      3
#> 506         1         2      3
#> 507         1         0      3
#> 508         1         0      3
#> 509         1         2      3
#> 510         1         2      3
#> 511         1         2      3
#> 512         0         0      3
#> 513         1         2      3
#> 514         1         0      3
#> 515         1         2      3
#> 516         1         0      3
#> 517         1         2      3
#> 518         1         0      3
#> 519         1         0      3
#> 520         1         2      3
#> 521         0         0      3
#> 522         0         2      3
#> 523         1         0      3
#> 524         1         2      3
#> 525         1         0      3
#> 526         1         2      3
#> 527         1         2      3
#> 528         1         2      3
#> 529         1         0      3
#> 530         1         0      3
#> 531         1         2      3
#> 532         0         0      3
#> 533         1         2      3
#> 534         1         0      3
#> 535         1         0      3
#> 536         1         2      3
#> 537         1         2      3
#> 538         0         0      3
#> 539         1         2      3
#> 540         1         2      3
#> 541         0         0      3
#> 542         1         0      3
#> 543         1         2      3
#> 544         1         0      3
#> 545         1         0      3
#> 546         1         2      3
#> 547         1         0      3
#> 548         0         0      3
#> 549         1         2      3
#> 550         1         2      3
#> 551         0         2      4
#> 552         1         0      4
#> 553         1         0      4
#> 554         0         2      4
#> 555         1         3      4
#> 556         1         3      4
#> 557         1         2      4
#> 558         1         3      4
#> 559         0         0      4
#> 560         1         3      4
#> 561         0         2      4
#> 562         0         0      4
#> 563         0         3      4
#> 564         1         0      4
#> 565         0         0      4
#> 566         1         3      4
#> 567         1         2      4
#> 568         0         2      4
#> 569         1         3      4
#> 570         0         3      4
#> 571         1         2      4
#> 572         1         0      4
#> 573         1         0      4
#> 574         0         2      4
#> 575         1         3      4
#> 576         0         3      4
#> 577         1         0      4
#> 578         1         2      4
#> 579         0         2      4
#> 580         1         0      4
#> 581         1         3      4
#> 582         0         2      4
#> 583         1         2      4
#> 584         0         0      4
#> 585         1         3      4
#> 586         1         0      4
#> 587         1         0      4
#> 588         1         3      4
#> 589         1         0      4
#> 590         0         2      4
#> 591         1         2      4
#> 592         1         3      4
#> 593         1         3      4
#> 594         1         0      4
#> 595         1         3      4
#> 596         1         0      4
#> 597         1         2      4
#> 598         1         2      4
#> 599         1         3      4
#> 600         1         2      4
#> 601         1         0      4
#> 602         1         3      4
#> 603         1         0      4
#> 604         1         2      4
#> 605         1         3      4
#> 606         1         0      4
#> 607         0         2      4
#> 608         1         2      4
#> 609         1         0      4
#> 610         0         3      4
#> 611         1         3      4
#> 612         1         3      4
#> 613         0         2      4
#> 614         0         0      4
#> 615         1         2      4
#> 616         0         0      4
#> 617         1         3      4
#> 618         1         0      4
#> 619         0         0      4
#> 620         0         3      4
#> 621         1         2      4
#> 622         1         2      4
#> 623         1         3      4
#> 624         1         0      4
#> 625         0         2      4
#> 626         1         2      4
#> 627         0         3      4
#> 628         1         0      4
#> 629         0         2      4
#> 630         1         0      4
#> 631         1         2      4
#> 632         1         0      4
#> 633         1         3      4
#> 634         1         3      4
#> 635         1         3      4
#> 636         1         0      4
#> 637         1         2      4
#> 638         1         0      4
#> 639         1         2      4
#> 640         1         3      4
#> 641         1         2      4
#> 642         1         3      4
#> 643         1         3      4
#> 644         1         0      4
#> 645         1         2      4
#> 646         0         0      4
#> 647         1         0      4
#> 648         1         3      4
#> 649         1         2      4
#> 650         1         0      4
#> 651         1         3      4
#> 652         1         2      4
#> 653         1         3      4
#> 654         0         0      4
#> 655         0         0      4
#> 656         1         2      4
#> 657         1         2      4
#> 658         0         3      4
#> 659         0         2      4
#> 660         1         0      4
#> 661         0         3      4
#> 662         1         0      4
#> 663         1         2      4
#> 664         1         3      4
#> 665         1         2      4
#> 666         1         3      4
#> 667         1         3      4
#> 668         1         2      4
#> 669         1         0      4
#> 670         1         0      4
#> 671         1         0      4
#> 672         0         3      4
#> 673         0         2      4
#> 674         1         2      4
#> 675         1         3      4
#> 676         1         0      4
#> 677         1         0      4
#> 678         0         0      4
#> 679         1         3      4
#> 680         1         3      4
#> 681         1         2      4
#> 682         1         2      4
#> 683         1         3      4
#> 684         1         0      4
#> 685         0         0      4
#> 686         1         2      4
#> 687         1         2      4
#> 688         1         3      4
#> 689         1         0      4
#> 690         1         3      4
#> 691         0         0      4
#> 692         0         2      4
#> 693         1         2      4
#> 694         1         3      4
#> 695         1         2      4
#> 696         1         3      4
#> 697         1         0      4
#> 698         1         0      4
#> 699         1         2      4
#> 700         1         3      4
#> 701         1         0      4
#> 702         1         2      4
#> 703         1         3      4
#> 704         0         2      4
#> 705         0         3      4
#> 706         0         0      4
#> 707         1         2      4
#> 708         1         2      4
#> 709         1         0      4
#> 710         1         0      4
#> 711         1         3      4
#> 712         0         3      4
#> 713         1         0      4
#> 714         0         2      4
#> 715         1         3      4
#> 716         1         2      4
#> 717         0         0      4
#> 718         0         3      4
#> 719         1         3      4
#> 720         1         0      4
#> 721         1         2      4
#> 722         1         0      4
#> 723         1         3      4
#> 724         1         2      4
#> 725         1         0      4
#> 726         1         3      4
#> 727         1         2      4
#> 728         1         0      4
#> 729         1         3      4
#> 730         1         2      4
#> 731         1         3      4
#> 732         1         0      4
#> 733         1         2      4
#> 734         1         0      4
#> 735         1         3      4
#> 736         0         2      4
#> 737         1         3      4
#> 738         1         0      4
#> 739         1         2      4
#> 740         1         3      4
#> 741         0         0      4
#> 742         1         2      4
#> 743         1         2      4
#> 744         1         3      4
#> 745         1         0      4
#> 746         1         2      4
#> 747         1         3      4
#> 748         1         0      4
#> 749         1         2      4
#> 750         1         3      4
#> 751         0         2      4
#> 752         0         0      4
#> 753         1         3      4
#> 754         1         0      4
#> 755         1         3      4
#> 756         1         0      4
#> 757         1         0      4
#> 758         0         3      4
#> 759         1         2      4
#> 760         1         2      4
#> 761         1         3      5
#> 762         1         0      5
#> 763         0         0      5
#> 764         1         3      5
#> 765         0         0      5
#> 766         0         0      5
#> 767         1         3      5
#> 768         1         3      5
#> 769         1         0      5
#> 770         1         3      5
#> 771         1         0      5
#> 772         1         3      5
#> 773         1         3      5
#> 774         1         3      5
#> 775         1         0      5
#> 776         1         0      5
#> 777         0         3      5
#> 778         0         3      5
#> 779         1         0      5
#> 780         0         0      5
#> 781         1         3      5
#> 782         1         0      5
#> 783         1         3      5
#> 784         1         0      5
#> 785         1         0      5
#> 786         1         3      5
#> 787         1         0      5
#> 788         1         3      5
#> 789         0         0      5
#> 790         0         3      5
#> 791         1         3      5
#> 792         1         0      5
#> 793         0         0      5
#> 794         1         3      5
#> 795         1         3      5
#> 796         1         0      5
#> 797         1         3      5
#> 798         1         0      5
#> 799         1         3      5
#> 800         0         0      5
#> 801         1         0      5
#> 802         1         3      5
#> 803         1         3      5
#> 804         1         0      5
#> 805         1         3      5
#> 806         1         0      5
#> 807         0         3      5
#> 808         1         0      5
#> 809         0         0      5
#> 810         1         0      5
#> 811         1         3      5
#> 812         1         3      5
#> 813         1         3      5
#> 814         1         3      5
#> 815         1         0      5
#> 816         1         0      5
#> 817         1         3      5
#> 818         1         0      5
#> 819         1         0      5
#> 820         0         3      5
#> 821         0         3      5
#> 822         0         3      5
#> 823         1         0      5
#> 824         1         0      5
#> 825         1         3      5
#> 826         0         0      5
#> 827         1         0      5
#> 828         1         3      5
#> 829         0         0      5
#> 830         1         3      5
#> 831         1         0      5
#> 832         1         3      5
#> 833         0         3      5
#> 834         1         3      5
#> 835         1         0      5
#> 836         1         0      5
#> 837         0         3      5
#> 838         1         0      5
#> 839         0         0      5
#> 840         0         3      5
#> 841         1         0      5
#> 842         0         0      5
#> 843         1         3      5
#> 844         1         3      5
#> 845         1         0      5
#> 846         1         0      5
#> 847         1         3      5
#> 848         1         3      5
#> 849         0         3      5
#> 850         1         0      5
#> 851         1         0      5
#> 852         1         3      5
#> 853         1         3      5
#> 854         1         0      5
#> 855         1         3      5
#> 856         1         0      5
#> 857         1         3      5
#> 858         1         0      5
#> 859         1         3      5
#> 860         0         0      5
#> 861         1         0      5
#> 862         1         0      5
#> 863         1         3      5
#> 864         1         3      5
#> 865         1         0      5
#> 866         1         3      5
#> 867         0         3      5
#> 868         1         0      5
#> 869         0         3      5
#> 870         1         0      5
#> 871         1         3      5
#> 872         0         0      5
#> 873         1         3      5
#> 874         1         0      5
#> 875         1         0      5
#> 876         1         3      5
#> 877         0         3      5
#> 878         1         0      5
#> 879         1         3      5
#> 880         0         0      5
#> 881         1         3      5
#> 882         1         0      5
#> 883         0         0      5
#> 884         1         3      5
#> 885         1         0      5
#> 886         0         3      5
#> 887         0         0      5
#> 888         1         3      5
#> 889         1         0      5
#> 890         1         3      5
#> 891         0         0      5
#> 892         1         3      5
#> 893         1         0      5
#> 894         0         3      5
#> 895         1         3      5
#> 896         0         0      5
#> 897         1         3      5
#> 898         1         0      5
#> 899         0         3      5
#> 900         0         0      5
#> 901         1         3      5
#> 902         0         0      5
#> 903         1         3      5
#> 904         1         0      5
#> 905         1         3      5
#> 906         0         0      5
#> 907         1         3      5
#> 908         0         0      5
#> 909         1         0      5
#> 910         0         0      5
#> 911         1         3      5
#> 912         1         3      5
#> 913         1         3      5
#> 914         0         3      5
#> 915         1         0      5
#> 916         0         0      5
#> 917         1         3      5
#> 918         1         0      5
#> 919         1         3      5
#> 920         1         0      5
#> 921         1         0      5
#> 922         1         0      5
#> 923         1         3      5
#> 924         1         3      5
#> 925         1         3      5
#> 926         1         0      5
#> 927         1         0      5
#> 928         1         3      5
#> 929         1         0      5
#> 930         1         3      5
#> 931         1         0      5
#> 932         1         3      5
#> 933         0         0      5
#> 934         1         0      5
#> 935         1         3      5
#> 936         0         3      5
#> 937         1         3      5
#> 938         0         0      5
#> 939         1         0      5
#> 940         1         3      5
#> 941         1         3      5
#> 942         1         3      5
#> 943         1         0      5
#> 944         1         0      5
#> 945         1         3      5
#> 946         1         0      5
#> 947         0         0      5
#> 948         1         3      5
#> 949         1         0      5
#> 950         1         0      5
#> 951         0         3      5
#> 952         1         3      5
#> 953         1         3      5
#> 954         1         0      5
#> 955         1         0      5
#> 956         1         3      5
#> 957         1         0      5
#> 958         1         0      5
#> 959         0         3      5
#> 960         1         3      5
#> 961         1         3      5
#> 962         1         0      5
#> 963         0         3      5
#> 964         1         0      5
#> 965         1         0      5
#> 966         1         3      5
#> 967         1         0      5
#> 968         1         3      5
#> 969         1         0      5
#> 970         0         0      5
#> 971         1         3      5
#> 972         1         3      5
#> 973         0         3      5
#> 974         0         0      5
#> 975         0         3      5
#> 976         1         0      5
#> 977         1         0      5
#> 978         0         3      5
#> 979         1         3      5
#> 980         0         0      5
#> 981         1         3      5
#> 982         1         3      5
#> 983         1         0      5
#> 984         1         0      5
#> 985         0         3      5
#> 986         1         3      5
#> 987         0         0      5
#> 988         1         0      5
#> 989         0         0      5
#> 990         1         3      5
#> 991         0         3      5
#> 992         1         0      5
#> 993         1         3      5
#> 994         1         3      5
#> 995         1         0      5
#> 996         1         0      5
#> 997         0         3      5
#> 998         1         0      5
#> 999         1         3      5
#> 1000        1         0      5

# Usage option 2 - specify n_arm and alloc_ratios
head(datasim_bin(n_total = NULL, num_arms = NULL, d = NULL,
n_arm = 200, alloc_ratios = matrix(c(1,1,1,1,1,NA,NA,1,1), ncol = 3, byrow = TRUE),
period_blocks = 2, p0 = 0.7, OR = rep(1.4, 3), lambda = rep(0.15, 4), trend = "linear"))
#>   response treatment period
#> 1        1         1      1
#> 2        1         0      1
#> 3        1         0      1
#> 4        1         1      1
#> 5        1         0      1
#> 6        1         1      1