Data simulation for binary endpoints for a platform trial with an arbitrary number of treatment arms entering sequentially

Simulates data from a platform trial with binary endpoints and an arbitrary number of treatment arms entering sequentially. There is the option to either specify the timing of adding arms and the overall sample size in the trial or specify the sample size per arm (assumed equal) and the allocation ratios per period.

Usage

datasim_bin(
  n_total,
  num_arms,
  d,
  n_arm,
  alloc_ratios,
  period_blocks = 2,
  p0,
  OR,
  lambda,
  trend,
  N_peak,
  full = FALSE
)

Arguments

n_total

Overall sample size for the trial

num_arms

Number of treatment arms in the trial

d

Timing of adding new arms in terms of number of patients allocated to the control arm

n_arm

Sample size per arm

alloc_ratios

Matrix with allocation ratios in each period (rows = arms; columns = periods)

period_blocks

Number to multiply the number of active arms with in order to get the block size per period

p0

Response in the control arm

OR

Vector with odds ratios for each arm

lambda

Vector with strength of time trend in each arm

trend

Indicates the time trend pattern ("linear", "stepwise" or "inv_u")

N_peak

Point at which the inverted-u time trend switches direction in terms of overall sample size

full

Boolean. Indicates whether only variables needed for the analysis should be in the output (FALSE) or also additional information (lambdas, underlying responses) should be included (TRUE). Default=FALSE

Value

Data frame: simulated trial data

Author

Pavla Krotka, Marta Bofill Roig

Examples


# Usage option 1 - specify n_total, num_arms and d
head(datasim_bin(n_total = 1000, num_arms = 3, d = 120,
n_arm = NULL, alloc_ratios = NULL,
period_blocks = 2, p0 = 0.7, OR = rep(1.4, 3), lambda = rep(0.15, 4), trend = "linear"))
#> Loading required package: rlang
#>   response treatment period
#> 1        0         0      1
#> 2        0         1      1
#> 3        1         0      1
#> 4        1         1      1
#> 5        1         1      1
#> 6        1         0      1

# Usage option 2 - specify n_arm and alloc_ratios
head(datasim_bin(n_total = NULL, num_arms = NULL, d = NULL,
n_arm = 200, alloc_ratios = matrix(c(1,1,1,1,1,NA,NA,1,1), ncol = 3, byrow = TRUE),
period_blocks = 2, p0 = 0.7, OR = rep(1.4, 3), lambda = rep(0.15, 4), trend = "linear"))
#>   response treatment period
#> 1        1         0      1
#> 2        1         1      1
#> 3        1         0      1
#> 4        1         1      1
#> 5        1         0      1
#> 6        1         1      1