Model-based analysis for continuous data using discontinuous piecewise polynomials per calendar time

Performs linear regression taking into account all trial data until the arm under study leaves the trial and adjusting for time using discontinuous piecewise polynomials in each calendar time interval

Usage

piecewise_cal_cont(
  data,
  arm,
  alpha = 0.025,
  unit_size = 25,
  ncc = TRUE,
  poly_degree = 3,
  ...
)

Arguments

data

Simulated trial data, e.g. result from the datasim_cont() function

arm

Indicator of the treatment arm under study to perform inference on (vector of length 1)

alpha

Type I error. Default=0.025

unit_size

Number of patients per calendar time unit Default=25

ncc

Boolean. Whether to include NCC data into the analysis. Default=TRUE

poly_degree

Degree of the piecewise polynomial. Default=3

...

Further arguments for simulation function

Value

List containing the p-value (one-sided), estimated treatment effect, 95% confidence interval, an indicator whether the null hypothesis was rejected or not for the investigated treatment and the fitted model

Author

Pavla Krotka

Examples


trial_data <- datasim_cont(num_arms = 3, n_arm = 100, d = c(0, 100, 250),
theta = rep(0.25, 3), lambda = rep(0.15, 4), sigma = 1, trend = "linear")

piecewise_cal_cont(data = trial_data, arm = 3)
#> $p_val
#> [1] 0.003675906
#> 
#> $treat_effect
#> [1] 0.3738407
#> 
#> $lower_ci
#> [1] 0.1010373
#> 
#> $upper_ci
#> [1] 0.6466442
#> 
#> $reject_h0
#> [1] TRUE
#> 
#> $model
#> 
#> Call:
#> lm(formula = response ~ as.factor(treatment) + poly(j, degree = poly_degree, 
#>     raw = TRUE) * as.factor(cal_time), data = data_new)
#> 
#> Coefficients:
#>                                                      (Intercept)  
#>                                                        1.116e+00  
#>                                            as.factor(treatment)1  
#>                                                        5.606e-01  
#>                                            as.factor(treatment)2  
#>                                                        2.663e-01  
#>                                            as.factor(treatment)3  
#>                                                        3.738e-01  
#>                       poly(j, degree = poly_degree, raw = TRUE)1  
#>                                                       -4.468e-01  
#>                       poly(j, degree = poly_degree, raw = TRUE)2  
#>                                                        3.572e-02  
#>                       poly(j, degree = poly_degree, raw = TRUE)3  
#>                                                       -7.752e-04  
#>                                             as.factor(cal_time)2  
#>                                                        5.342e+00  
#>                                             as.factor(cal_time)3  
#>                                                       -1.370e+02  
#>                                             as.factor(cal_time)4  
#>                                                       -1.207e+02  
#>                                             as.factor(cal_time)5  
#>                                                        2.259e+01  
#>                                             as.factor(cal_time)6  
#>                                                        1.602e+03  
#>                                             as.factor(cal_time)7  
#>                                                       -3.634e+03  
#>                                             as.factor(cal_time)8  
#>                                                        7.589e+02  
#>                                             as.factor(cal_time)9  
#>                                                        1.487e+03  
#>                                            as.factor(cal_time)10  
#>                                                        2.121e+03  
#>                                            as.factor(cal_time)11  
#>                                                       -2.087e+03  
#>                                            as.factor(cal_time)12  
#>                                                       -4.628e+03  
#>                                            as.factor(cal_time)13  
#>                                                       -2.165e+04  
#>                                            as.factor(cal_time)14  
#>                                                       -5.243e+03  
#>                                            as.factor(cal_time)15  
#>                                                       -3.367e+04  
#>                                            as.factor(cal_time)16  
#>                                                       -1.744e+03  
#>                                            as.factor(cal_time)17  
#>                                                       -4.917e+04  
#>                                            as.factor(cal_time)18  
#>                                                        9.232e+04  
#>                                            as.factor(cal_time)19  
#>                                                       -3.298e+04  
#>                                            as.factor(cal_time)20  
#>                                                       -1.162e+04  
#>  poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)2  
#>                                                        1.438e-01  
#>  poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)2  
#>                                                       -3.397e-02  
#>  poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)2  
#>                                                        8.078e-04  
#>  poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)3  
#>                                                        7.284e+00  
#>  poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)3  
#>                                                       -1.501e-01  
#>  poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)3  
#>                                                        1.410e-03  
#>  poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)4  
#>                                                        4.463e+00  
#>  poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)4  
#>                                                       -8.031e-02  
#>  poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)4  
#>                                                        9.388e-04  
#>  poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)5  
#>                                                       -1.112e+00  
#>  poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)5  
#>                                                       -1.364e-02  
#>  poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)5  
#>                                                        6.859e-04  
#>  poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)6  
#>                                                       -3.481e+01  
#>  poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)6  
#>                                                        2.223e-01  
#>  poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)6  
#>                                                        1.471e-04  
#>  poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)7  
#>                                                        6.704e+01  
#>  poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)7  
#>                                                       -4.422e-01  
#>  poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)7  
#>                                                        1.601e-03  
#>  poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)8  
#>                                                       -1.136e+01  
#>  poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)8  
#>                                                        2.571e-02  
#>  poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)8  
#>                                                        6.682e-04  
#>  poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)9  
#>                                                       -2.070e+01  
#>  poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)9  
#>                                                        6.441e-02  
#>  poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)9  
#>                                                        6.172e-04  
#> poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)10  
#>                                                       -2.727e+01  
#> poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)10  
#>                                                        8.462e-02  
#> poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)10  
#>                                                        6.015e-04  
#> poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)11  
#>                                                        2.477e+01  
#> poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)11  
#>                                                       -1.302e-01  
#> poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)11  
#>                                                        8.974e-04  
#> poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)12  
#>                                                        5.033e+01  
#> poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)12  
#>                                                       -2.148e-01  
#> poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)12  
#>                                                        9.894e-04  
#> poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)13  
#>                                                        2.057e+02  
#> poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)13  
#>                                                       -6.844e-01  
#> poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)13  
#>                                                        1.458e-03  
#> poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)14  
#>                                                        4.912e+01  
#> poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)14  
#>                                                       -1.861e-01  
#> poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)14  
#>                                                        9.299e-04  
#> poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)15  
#>                                                        2.782e+02  
#> poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)15  
#>                                                       -7.992e-01  
#> poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)15  
#>                                                        1.475e-03  
#> poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)16  
#>                                                        1.378e+01  
#> poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)16  
#>                                                       -6.964e-02  
#> poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)16  
#>                                                        8.039e-04  
#> poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)17  
#>                                                        3.567e+02  
#> poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)17  
#>                                                       -8.962e-01  
#> poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)17  
#>                                                        1.468e-03  
#> poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)18  
#>                                                       -6.309e+02  
#> poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)18  
#>                                                        1.403e+00  
#> poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)18  
#>                                                       -3.177e-04  
#> poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)19  
#>                                                        2.118e+02  
#> poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)19  
#>                                                       -4.870e-01  
#> poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)19  
#>                                                        1.097e-03  
#> poly(j, degree = poly_degree, raw = TRUE)1:as.factor(cal_time)20  
#>                                                        6.620e+01  
#> poly(j, degree = poly_degree, raw = TRUE)2:as.factor(cal_time)20  
#>                                                       -1.588e-01  
#> poly(j, degree = poly_degree, raw = TRUE)3:as.factor(cal_time)20  
#>                                                        8.513e-04  
#> 
#>